split filesé
